VC documents and links BoK - refactor the AddContent
